+++ cf/data/creation.py
@@ -114,7 +114,7 @@ def to_dask(array, chunks, **from_array_options):
 
     try:
         return da.from_array(array, chunks=chunks, **kwargs)
-    except NotImplementedError as error:
+    except NotImplementedError:
         # Try again with 'chunks=-1', in case the failure was due to
         # not being able to use auto rechunking with object dtype.
         return da.from_array(array, chunks=-1, **kwargs)